requirements.txt
langchain
tiktoken==0.3.2
ratelimiter
python-dotenv==1.0.0
markdownify==0.11.6
openai
faiss-cpu
streamlit
streamlit-chat